# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Syncattend (SDAS) API Contract
  version: 0.1.0-draft
  description: >
    Shared API contract for Syncattend V3 (dual-auth smart attendance).
    OWNERSHIP: this file is the single source of truth for the HTTP contract.
    Only backend (owner A) may modify it. mobile (B) and web (C) READ ONLY.
    Layered defense: dynamic QR (time/identity) + ultrasonic audio token
    (space/proximity) + identity defense (1-min window + server uniqueness +
    app-generated UUID device binding + @wku.ac.kr re-registration) + professor-in-the-loop.
servers:
  - url: http://localhost:8000
    description: Local dev (backend)
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new account (student/professor)
      description: >
        Creates an account with a bcrypt/argon2-hashed password. Student
        accounts must use a @wku.ac.kr email (identity defense). Duplicate
        emails are rejected with 409.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterUserRequest' }
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserOut' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: Email already registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/login:
    post:
      tags: [auth]
      summary: Login, issue JWT
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: JWT issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ------------------------------------------------------------- devices
...
  /attendance/verify:
    post:
      tags: [attendance]
      summary: Student cross-verification (window + QR×audio + single-use nonce + device + uniqueness)
      description: >
        Server verification order:
        (1) auth window open, (2) QR×audio cross-verify,
        (3) single-use nonce not yet consumed (consume on pass),
        (4) device_uuid matches account binding,
        (5) uniqueness: one attendance per account & per device_uuid per session.
      operationId: verifyAttendance
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyRequest' }
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerifyResult' }
        '409':
          description: Rejected (see reason)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerifyResult' }
  /sessions/{id}/attendance:
    get:
      tags: [attendance]
      summary: Live attendance rate + unverified list (professor)
      operationId: getSessionAttendance
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      responses:
        '200':
          description: Aggregation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttendanceAggregate' }
  /sessions/{id}/attendance/batch:
    post:
      tags: [attendance]
      summary: Delta-based batch close (only changed/unverified records)
      operationId: batchCloseAttendance
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deltas]
              properties:
                deltas:
                  type: array
                  items: { $ref: '#/components/schemas/AttendanceDelta' }
      responses:
        '200':
          description: Batch applied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttendanceAggregate' }
  /attendance/{record_id}:
    patch:
      tags: [attendance]
      summary: Manual correction of one record (professor-in-the-loop)
      operationId: correctAttendance
      security: [{ bearerAuth: [] }]
      parameters:
        - name: record_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/AttendanceStatus' }
      responses:
        '200':
          description: Updated record
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttendanceRecord' }

  # ---------------------------------------------------------------- sse
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      properties:
        detail: { type: string }
        code: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: bearer }
        role: { $ref: '#/components/schemas/Role' }
    Role:
      type: string
      enum: [student, professor]
    RegisterUserRequest:
      type: object
      required: [email, password, role]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        role: { $ref: '#/components/schemas/Role' }
        name: { type: string, nullable: true }
    UserOut:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        name: { type: string, nullable: true }
...
    VerifyRequest:
      type: object
      required: [session_id, qr_token, audio_nonce, device_uuid]
      properties:
        session_id: { type: string }
        qr_token: { type: string }
        audio_nonce: { type: string }
        device_uuid: { type: string, format: uuid }
    VerifyResult:
      type: object
      properties:
        status: { type: string, enum: [present, pending, rejected] }
        cross_verified: { type: boolean }
        device_matched: { type: boolean }
        reason:
          type: string
          enum: [window_closed, cross_verify_failed, nonce_reused, device_mismatch, duplicate_attendance]
          nullable: true
        risk_warning:
          $ref: '#/components/schemas/RiskWarning'
          nullable: true
    AttendanceStatus:
      type: string
      enum: [present, absent, pending]
    AttendanceRecord:
      type: object
      properties:
        record_id: { type: string }
        session_id: { type: string }
        student_id: { type: string }
        status: { $ref: '#/components/schemas/AttendanceStatus' }
        verified_at: { type: string, format: date-time, nullable: true }
    AttendanceDelta:
      type: object
      required: [student_id, status]
      properties:
        student_id: { type: string }
        status: { $ref: '#/components/schemas/AttendanceStatus' }
    AttendanceAggregate:
      type: object
      properties:
        session_id: { type: string }
        total: { type: integer }
        present: { type: integer }
        unverified:
          type: array
          items: { type: string, description: student_id }
    RiskWarning:
      type: object
      properties:
        level: { type: string, enum: [info, warning, danger] }
        absences: { type: integer }
        message: { type: string }